Conflicts:
	Sampoerna.EMS.DAL/EMSEntities.edmx

+++ Sampoerna.EMS.Website/Models/PBCK1/PBCK1ProdPlanModel.cs
@@ -3,7 +3,7 @@
     public class Pbck1ProdPlanModel
     {
         public long Pbck1ProdPlanId { get; set; }
-        public long Pbck1Id { get; set; }
+        public int Pbck1Id { get; set; }
         public string ProductCode { get; set; }
         public string ProdTypeName { get; set; }
         public string ProdTypeAlias { get; set; }
